tests/mse_mp4/starlight-init.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: load init segment rdar://77785362</title>
  <script type="text/javascript" src="mediasource.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

runWithMSE(function(ms, el) {
  el.controls = true;
  once(ms, 'sourceopen').then(function() {
    // Log events for debugging.
    var events = ["suspend", "play", "canplay", "canplaythrough", "loadstart", "loadedmetadata",
                  "loadeddata", "playing", "ended", "error", "stalled", "emptied", "abort",
                  "waiting", "pause", "durationchange", "seeking", "seeked"];
    function logEvent(e) {
      var v = e.target;
      info("got " + e.type + " event");
    }
    events.forEach(function(e) {
      el.addEventListener(e, logEvent);
    });

    ok(true, "Receive a sourceopen event");
    var audiosb = ms.addSourceBuffer('audio/mp4; codecs="mp4a.40.2"');
    el.addEventListener("error", function(e) {
      ok(false, "should not fire '" + e.type + "' event");
      SimpleTest.finish();
    });
    is(el.readyState, el.HAVE_NOTHING, "readyState is HAVE_NOTHING");
    var promises = [];
    promises.push(fetchAndLoad(audiosb, 'starlight-', ['init'], '.mp4'));
    promises.push(once(el, 'loadedmetadata'));
    Promise.all(promises)
    .then(function() {
      ok(true, "got loadedmetadata event");
      var promises = [];
      ms.endOfStream();
      return once(ms, 'sourceended');
    })
    .then(function() {
      SimpleTest.finish();
    })
  });
});

</script>
</pre>
</body>
</html>
